Printf

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: tmoragli <tmoragli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/30 15:37:43 by telli             #+#    #+#             */
-/*   Updated: 2021/01/27 15:53:08 by tmoragli         ###   ########.fr       */
+/*   Updated: 2021/07/18 16:57:06 by tmoragli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
-
-size_t		ft_strlen(const char *str)
+#include "printf.h"
+long int	ft_strlen(const char *str)
 {
-	size_t i;
+	long int i;
 
 	i = 0;
 	while (str[i])
